Show download progress with human-readable units

+++ src/main.rs
@@ -95,7 +95,7 @@ async fn main() -> Result<()> {
 
     let attachments = jira.fetch_attachments(&args.issue).await?;
     for att in &attachments {
-        let formatted_size = app::format_file_size(att.size as usize);
+        let formatted_size = app::format_file_size(att.size);
         info!(
             "Attachment: \"{}\" ({}) - {}",
             att.filename, formatted_size, att.created
